requirements.txt
torch==2.1.0
transformers==4.36.0
peft==0.7.1
scikit-learn==1.3.2
iterative-stratification==0.1.7
pandas==2.1.4
numpy==1.26.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
tensorboard==2.15.1
pyyaml==6.0.1
sentence-transformers==2.2.2
